# modules/fundamental/src/vulnerability/endpoints/mod.rs
#[cfg(test)]
mod test;

use crate::{
    Error::{self},
    endpoints::Deprecation,
    vulnerability::{
        model::{AnalysisRequest, AnalysisResponse, VulnerabilityDetails, VulnerabilitySummary},
        service::VulnerabilityService,
    },
};
use actix_web::{HttpResponse, Responder, delete, get, post, web};
use trustify_auth::{DeleteVulnerability, ReadAdvisory, authorizer::Require};
use trustify_common::{
    db::{Database, query::Query},
    model::{Paginated, PaginatedResults},
};

pub fn configure(config: &mut utoipa_actix_web::service_config::ServiceConfig, db: Database) {
    let service = VulnerabilityService::new();
    config
        .app_data(web::Data::new(service))
        .app_data(web::Data::new(db))
        .service(all)
        .service(delete)
        .service(get)
        .service(analyze);
}

#[utoipa::path(
    tag = "vulnerability",
    operation_id = "listVulnerabilities",
    params(
        Query,
        Paginated,
    ),
    responses(
        (status = 200, description = "Matching vulnerabilities", body = PaginatedResults<VulnerabilitySummary>),
    ),
)]
#[get("/v2/vulnerability")]
/// List vulnerabilities
pub async fn all(
    state: web::Data<VulnerabilityService>,
    db: web::Data<Database>,
    web::Query(search): web::Query<Query>,
    web::Query(paginated): web::Query<Paginated>,
    web::Query(Deprecation { deprecated }): web::Query<Deprecation>,
    _: Require<ReadAdvisory>,
) -> actix_web::Result<impl Responder> {
    Ok(HttpResponse::Ok().json(
        state
            .fetch_vulnerabilities(search, paginated, deprecated, db.as_ref())
            .await?,
    ))
}

#[utoipa::path(
    tag = "vulnerability",
    operation_id = "getVulnerability",
    params(
        ("id", Path, description = "ID of the vulnerability")
    ),
    responses(
        (status = 200, description = "Specified vulnerability", body = VulnerabilityDetails),
        (status = 404, description = "The vulnerability could not be found"),
    ),
)]
#[get("/v2/vulnerability/{id}")]
/// Retrieve vulnerability details
pub async fn get(
    state: web::Data<VulnerabilityService>,
    db: web::Data<Database>,
    id: web::Path<String>,
    web::Query(Deprecation { deprecated }): web::Query<Deprecation>,
    _: Require<ReadAdvisory>,
) -> actix_web::Result<impl Responder> {
    let vuln = state
        .fetch_vulnerability(&id, deprecated, db.as_ref())
        .await?;
    if let Some(vuln) = vuln {
        Ok(HttpResponse::Ok().json(vuln))
    } else {
        Ok(HttpResponse::NotFound().finish())
    }
}

#[utoipa::path(
    tag = "vulnerability",
    operation_id = "deleteVulnerability",
    params(
        ("id", Path, description = "ID of the vulnerability")
    ),
    responses(
        (status = 501, description = "Method not yet implemented"),
    ),
)]

#[utoipa::path(
  operation_id = "analyze",
  tag = "vulnerability",
  request_body = AnalysisRequest,
  responses(
      (status = 200, description = "Analyze the provided purls to search for known vulnerabilities", body = AnalysisResponse),
  ),
)]
#[post("/v2/vulnerability/analyze")]
/// Analyze the provided purls for the known vulnerabilities
pub async fn analyze(
    service: web::Data<VulnerabilityService>,
    db: web::Data<Database>,
    web::Json(AnalysisRequest { purls }): web::Json<AnalysisRequest>,
    _: Require<ReadAdvisory>,
) -> actix_web::Result<impl Responder> {
    let details = service.analyze_purls(purls, db.as_ref()).await?;

    Ok(HttpResponse::Ok().json(details))
}